01-rag-document-q-n-a/requirements.txt
# python 3.12.9
#
streamlit==1.35.0
streamlit-option-menu ==0.3.12
streamlit-lottie==0.0.5
streamlit-pandas-profiling==0.1.3
streamlit-pdf-viewer==0.0.14
streamlit_antd_components==0.3.2
langchain==0.3.19
langchain-community==0.3.18
PyPDF2==3.0.1
python-dotenv==1.0.0
tiktoken==0.9.0
faiss-cpu==1.10.0
openai==1.65.2
